+++ src/components/Nav/Nav.js
@@ -1,16 +1,23 @@
 //Required Libraries
 import React from "react";
 
+//Required Components
+import SettingsMenu from "../Popups/SettingsMenu";
+
 /*
 The Nav consists of the following: 
 - Toggle between YouTube and Spotify (Video / Audio)
 - Drop down for Playlist Settings (Edit, Create, Delete)
 - Drop down for selecting current playlist
 */
-const Nav = () => {
+const Nav = (props) => {
     return(
         <React.Fragment>
-            <h1>Nav</h1>
+            <div className="center">
+                <h1>-Nav-</h1>
+                <h1 onClick={() => {props.setPopUp(<SettingsMenu setPopUp={props.setPopUp}/>)}}>-Settings-</h1>
+            </div>
+            
         </React.Fragment>
     )
 }