+++ src/components/Popups/MediaInstance.js
@@ -1,0 +1,1 @@
+//This component is the button you press to add a media to a playlist.