+++ src/components/Popups/PlaylistSettings.js
@@ -1,0 +1,18 @@
+//This component allows users to create and save playlists as well as change the playlist that is being edited. 
+//required libraries
+import React from "react";
+
+const PlaylistSettings = () => {
+
+
+
+    return(    
+        <div className="settings-grid-playlist-settings center flex-column">
+            <h1>Playlist Settings</h1>
+            
+            
+        </div>
+        )
+}
+
+export default PlaylistSettings;